make gossip ref tests nicely structured
