+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/teku/reference/common/gossip/GossipAttesterSlashingTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.reference.fulu.networking;
+package tech.pegasys.teku.reference.common.gossip;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static tech.pegasys.teku.reference.TestDataUtils.loadSsz;
