+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/teku/reference/fulu/networking/GetCustodyGroupsTestExecutor.java
@@ -14,55 +14,55 @@
 package tech.pegasys.teku.reference.fulu.networking;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static tech.pegasys.teku.reference.TestDataUtils.loadYaml;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.io.IOException;
 import java.math.BigInteger;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 import org.apache.tuweni.units.bigints.UInt256;
 import tech.pegasys.teku.ethtests.finder.TestDefinition;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.reference.TestDataUtils;
 import tech.pegasys.teku.reference.TestExecutor;
-import tech.pegasys.teku.spec.SpecVersion;
 import tech.pegasys.teku.spec.logic.versions.fulu.helpers.MiscHelpersFulu;
 
 public class GetCustodyGroupsTestExecutor implements TestExecutor {
 
   @Override
-  public void runTest(final TestDefinition testDefinition) throws Exception {
-    final GetCustodyGroupsMetaData metaData =
-        loadYaml(testDefinition, "meta.yaml", GetCustodyGroupsMetaData.class);
-    final SpecVersion spec = testDefinition.getSpec().getGenesisSpec();
-    final List<UInt64> actualResult =
-        MiscHelpersFulu.required(spec.miscHelpers())
-            .getCustodyGroups(metaData.getNodeId(), metaData.getCustodyGroupCount());
-    assertThat(new HashSet<>(actualResult)).isEqualTo(metaData.getResult());
+  public void runTest(final TestDefinition testDefinition) throws Throwable {
+    final MiscHelpersFulu miscHelpersFulu =
+        MiscHelpersFulu.required(testDefinition.getSpec().getGenesisSpec().miscHelpers());
+    final Data data = loadDataFile(testDefinition, Data.class);
+
+    final List<UInt64> calculatedCustodyGroups =
+        miscHelpersFulu.getCustodyGroups(data.nodeId(), data.custodyGroupCount);
+
+    assertThat(calculatedCustodyGroups).isEqualTo(data.expectedCustodyGroups());
   }
 
-  private static class GetCustodyGroupsMetaData {
+  private static class Data {
 
     @JsonProperty(value = "node_id", required = true)
     private String nodeId;
 
     @JsonProperty(value = "custody_group_count", required = true)
-    private int custodyGroupCount;
+    private Integer custodyGroupCount;
 
     @JsonProperty(value = "result", required = true)
     private List<Integer> result;
 
-    public UInt256 getNodeId() {
-      return UInt256.valueOf(new BigInteger(nodeId));
+    public UInt256 nodeId() {
+      return UInt256.valueOf(new BigInteger(this.nodeId, 10));
     }
 
-    public int getCustodyGroupCount() {
-      return custodyGroupCount;
+    public List<UInt64> expectedCustodyGroups() {
+      return result.stream().map(UInt64::valueOf).collect(Collectors.toList());
     }
+  }
 
-    public Set<UInt64> getResult() {
-      return result.stream().map(UInt64::valueOf).collect(Collectors.toUnmodifiableSet());
-    }
+  protected <T> T loadDataFile(final TestDefinition testDefinition, final Class<T> type)
+      throws IOException {
+    return TestDataUtils.loadYaml(testDefinition, "meta.yaml", type);
   }
 }
